mapper/requirements.txt
Django==1.8.3
djangorestframework==3.1.3
psycopg2==2.6.1
wsgiref==0.1.2
